+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Support for high DPI for WPF (#149)
 - EdgeRenderingMode property to PlotElement, allowing customization of the way edges are treated by the renderer (#1428, #1358, #1077, #843, #145)
 - Color palettes Viridis, Plasma, Magma, Inferno and Cividis (#1505)
+- Renderer based on SkiaSharp, including exporters for PNG, JPEG, PDF and SVG (#1509)
 
 ### Changed
 - Legends model (#644)